SF: Scheduler: set an expiration for heuristic layer's high refresh rate usages
Non-interactive heuristic layer's high refresh rate usages are overkill
and unnecessary.

Set a hard expiration limit to 2 seconds.

This effectively fixes/workarounds a major shortcoming in the whole
scheduler where it's totally unaware of the view's sizes.

For example, Instagram Reels with audios are played with a full-screen
video layer and a small animating audio icon at the bottom.

Because the audio icon's animations are lightweight, it's done at the
full, highest possible refresh rate. This prevents the device from
entering lower refresh rate states regardless of the video's fps.

With this commit, such layers are not allowed to use > 60 Hz states for
longer than 2 seconds. Note that this is separate from touch boost.

[ghostrider-reborn: adapt to 13]

Change-Id: I065250bbbe3f0b9ff071779bcc43a4499463ef24
Co-authored-by: Adithya R <gh0strider.2k18.reborn@gmail.com>
Signed-off-by: Dmitrii <bankersenator@gmail.com>